feat: move free trial out of assistant
